+++ config/passport.js
@@ -1,7 +1,7 @@
 const passport = require('passport');
 const GoogleStrategy = require('passport-google-oauth').OAuth2Strategy;
 const users = require("../models/User");
-const videos = require("../models/videos");
+
 
 passport.use(new GoogleStrategy({ 
       clientID: process.env.Google_Client_ID,
